Add functions to despike polarization sequences of images.
The basic idea is that when fed a sequence of images with the same
pointing, to look for spikes at a given pixel that are well outside
the norm for that pixel based on the other images in the sequence.
The actual comparison is done on a high-pass-filtered version of
the images, so that the detection works well in bright regions
as well as less bright regions.